requirements.txt
fpdf2==2.8.1
google-generativeai==0.8.3
grpcio-status==1.67.0
openai==1.52.2
pdf2image==1.17.0
poppler-utils==0.1.0
PyMuPDF==1.24.12
reflex-img-comparison-slider==0.0.1
replicate==0.26.0
ultralytics==8.3.23
reportlab>=4.0.0
Pillow==11.0.0
